not again

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -5,16 +5,15 @@
  * @n: int type number
  *
  * Description: Can only use _putchar to print.
- *
  */
 
 void print_number(int n)
 {
-	long m; /* power of 10 */
+	int m; /* power of 10 */
 
 	int c; /* boolean check */
 
-	long num; /* convert int to long */
+	int num; /* convert int to long */
 
 	num = n;
